Manually merge in changes from upstream upstream

+++ src/MO/src/modinfodialog.cpp
@@ -333,7 +333,7 @@ void ModInfoDialog::refreshLists() {
 
 void ModInfoDialog::addCategories(const CategoryFactory& factory, const std::set<int>& enabledCategories,
                                   QTreeWidgetItem* root, int rootLevel) {
-    for (size_t i = 0; i < factory.numCategories(); ++i) {
+    for (int i = 0; i < static_cast<int>(factory.numCategories()); ++i) {
         if (factory.getParentID(i) != rootLevel) {
             continue;
         }
